refactor(agent-v2): localize web app launch action
